Muneeb

+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/domain/Account.java
@@ -17,7 +17,7 @@ public class Account implements Serializable {
     private Integer accountId;
     private  String name;
     private  String description;
-    private Integer balance;
+    private double balance;
     @Enumerated(EnumType.STRING)
     AccountType accountType;
     @ManyToMany
